requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
PyPDF2==3.0.1
reportlab==4.0.7
weasyprint==60.2
openai==1.3.7
langchain==0.0.350
celery==5.3.4
redis==5.0.1
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4